fix(windows): remove all unused header file

+++ Engine/src/Platform/OpenGL/OpenGraphicsLibrary.cpp
@@ -91,7 +91,7 @@ namespace Engine
 		double xPos, yPos;
 		glfwGetCursorPos(m_window, &xPos, &yPos);
 
-		return { xPos, yPos };
+		return { (float)xPos, (float)yPos };
 	}
 
 	CursorMode OpenGraphicsLibrary::GetCursorMode() const
